signatures: Reorganize private modules
Split the `functions` module an put the type declarations with the
functions that use them.
