+++ crates/ruma-signatures/src/lib.rs
@@ -64,21 +64,19 @@ pub use ruma_common::{IdParseError, SigningKeyAlgorithm};
 pub use self::{
     ed25519::{Ed25519KeyPair, Ed25519KeyPairParseError, Ed25519VerificationError},
     error::{JsonError, VerificationError},
-    functions::{
-        canonical_json, content_hash, hash_and_sign_event, reference_hash, sign_json,
-        verify_canonical_json_bytes, verify_event, verify_json,
+    hash::{content_hash, reference_hash},
+    sign::{KeyPair, Signature, hash_and_sign_event, sign_json},
+    verify::{
+        PublicKeyMap, PublicKeySet, Verified, canonical_json, verify_canonical_json_bytes,
+        verify_event, verify_json,
     },
-    keys::{KeyPair, PublicKeyMap, PublicKeySet},
-    signatures::Signature,
-    verification::Verified,
 };
 
 mod ed25519;
 mod error;
-mod functions;
-mod keys;
-mod signatures;
-mod verification;
+mod hash;
+mod sign;
+mod verify;
 
 #[cfg(test)]
 mod tests {
